abstracted components into subcomponent
