+++ constants/types/index.ts
@@ -1,0 +1,3 @@
+import { MessageBoxProps } from './messageBox'
+
+export type { MessageBoxProps }
