--- lib/constants.ts
@@ -1,1 +1,0 @@
-export const WEBSITE_URL = 'https://nim-fawn.vercel.app'
